add "vagrant" command to build targets via Vagrant VM
This will launch a VM, rsync code into it, build the
dependent libraries via conan and finally compile qt
along with wkhtmltopdf and package it so that the
package is copied back into the host. This is done via
a custom Python module which implements the pre-build
and post-build steps.