use qemu-user-static which works with foreign docker images
The earlier multiarch/qemu-user-static:register image required the
right qemu-user-static binary be installed in the target image. The
fkrull/qemu-user-static image uses the "F - fix binary" flag so that
it works even within a container and hence works with vanilla images
for foreign architectures.

The minimum kernel requirement for emulation is now Linux 4.8.