leverage docker multi-arch support to simplify existing configuration
This is done via a new "matrix" key, which automatically generates
new targets for each supported architecture. We have an explicit
mapping from distro architecture => docker platform, as the naming
conventions vary wildly.

Also switch to a new image for qemu: https://github.com/dbhi/qus
and make sure we enable it only if it's really required. Have to
always use the 32-bit version due to a bug with glibc 2.28+, see
https://bugs.launchpad.net/qemu/+bug/1805913 for details.